SWER-88: plugin uninstall tweaks

+++ src/Migration/Migration1690444403DeleteLegacyCursors.php
@@ -21,19 +21,23 @@ class Migration1690444403DeleteLegacyCursors extends MigrationStep
 
     public function update(Connection $connection): void
     {
-        $connection->executeStatement(
-            'DELETE FROM `ergonode_cursor` where query in (:types)',
-            [
-                'types' => [
-                    CategoryStreamResultsProxy::MAIN_FIELD,
-                    CategoryTreeStreamResultsProxy::MAIN_FIELD,
-                    CategoryTreeStreamResultsProxy::TREE_LEAF_LIST_CURSOR,
+        $result = $connection->fetchOne('SHOW TABLES LIKE \'ergonode_cursor\';');
+
+        if ($result !== false) {
+            $connection->executeStatement(
+                'DELETE FROM `ergonode_cursor` where query in (:types)',
+                [
+                    'types' => [
+                        CategoryStreamResultsProxy::MAIN_FIELD,
+                        CategoryTreeStreamResultsProxy::MAIN_FIELD,
+                        CategoryTreeStreamResultsProxy::TREE_LEAF_LIST_CURSOR,
+                    ],
                 ],
-            ],
-            [
-                'types' => ArrayParameterType::STRING,
-            ]
-        );
+                [
+                    'types' => ArrayParameterType::STRING,
+                ]
+            );
+        }
     }
 
     public function updateDestructive(Connection $connection): void
